Add function and change structure
